species argument and switch argument species
The previous match.arg(species) was a type because it was missing the second argument to match from the catalog.  so this fix requires a user to enter geneWiseAnalysis(kexp, ... species="Homo.sapiens")  ; user must specify species of interest.

this line  species <- match.arg(species, c("Homo.sapiens",
                                       "Mus.musculus",
                                       "Rattus.norvegicus"))

will then yank out the species.  *The technically correct way of doing this is with a hash table*

Former-commit-id: a431d9d98afd3a2119f1c717c5d3d01fec13797b