Output limma expression values, with Meta info
Here is a summary that was tested on arcolombo/artemis  RAM branch.  this worked for the beta- version testing.  The motivation was due to Kwasi's request of having the differential expression with the meta-information.  this required data handling between features of genomic ranges, and limma outputs.


the issue is stemming from an early development decision to keep outputs in terms of ensembl gene IDs.  In order to use Reactome, this requires a mapping via biomaRt in terms of entrez id.    

Summary:
lines  81-101  perform   ensemblGene_ID -> entrez id  dependent on commonName derived from match.arg input from user.  this is standard and performed successfully from arcolombo/artemis ~ RAM. 
    -reactome is performed on the object converted[,2] which is all the non - na  entrez Ids.

line 120  runs a scaled transformation of limma values, but these values are in terms of ensembl ID,  so ensemblVector, are all the corresponding entries with non-na  entrez ID.  


lines 128-132  is over-coding in an attempt to protect the use of the "which" command.  here is the logic:
                      the ensemblVector is the first column of a data frame which stores all the non-NA entrez ID mapped by biomaRt.  
                      each column ensemblVector and entrezVector has identical rows.  
                      limma is performed on ensemblVector created res$scaledExprs and outputs ensembl IDs
                      line 123 converts the limma outputs to entrez with non-NA.
                      line 146 creates G_list that has human gene nomenclature
                      lines 156-181  create a merge between limma expression and meta information from features(kexp).

Lines 156-181 could be omitted if deemed unnecessary.  However lines 81-141 perform reactome and necessary ensembl -> entrez mapping.
Former-commit-id: 454e5a4fad83e97366af01ef9095323be5ddfce6